Preparing Portfolio
Add background for navigation menu, hover effects for the links and fixed burger with JavaScript to show and hide the menu on phones.

Animation for main title and navigation manu included.

Style the packs with images and a bit of information, same in info and mothership section.

Fixed the backgrounds and make them look better.

Mandela Samboy.